Replaced dependency from module Generic to module Common.

+++ data/scripts/upgrade.php
@@ -2,7 +2,7 @@
 
 namespace Zip;
 
-use Omeka\Stdlib\Message;
+use Common\Stdlib\PsrMessage;
 
 /**
  * @var Module $this
@@ -12,21 +12,33 @@ use Omeka\Stdlib\Message;
  *
  * @var \Omeka\Api\Manager $api
  * @var \Omeka\Settings\Settings $settings
+ * @var \Laminas\I18n\View\Helper\Translate $translate
  * @var \Doctrine\DBAL\Connection $connection
  * @var \Doctrine\ORM\EntityManager $entityManager
  * @var \Omeka\Mvc\Controller\Plugin\Messenger $messenger
  */
 $plugins = $services->get('ControllerPluginManager');
 $api = $plugins->get('api');
+// $config = $services->get('Config');
 $settings = $services->get('Omeka\Settings');
+$translate = $plugins->get('translate');
+// $translator = $services->get('MvcTranslator');
 $connection = $services->get('Omeka\Connection');
 $messenger = $plugins->get('messenger');
 $entityManager = $services->get('Omeka\EntityManager');
 
+if (!method_exists($this, 'checkModuleActiveVersion') || !$this->checkModuleActiveVersion('Common', '3.4.65')) {
+    $message = new \Omeka\Stdlib\Message(
+        $translate('The module %1$s should be upgraded to version %2$s or later.'), // @translate
+        'Common', '3.4.65'
+    );
+    throw new \Omeka\Module\Exception\ModuleCannotInstallException((string) $message);
+}
+
 if (version_compare($oldVersion, '3.4.0.3', '<')) {
     $settings->delete('zip_job');
 
-    $message = new Message(
+    $message = new PsrMessage(
         'It is now possible to store resources separately with files.' // @translate
     );
     $messenger->addSuccess($message);
